Safari-611.1.10/LayoutTests/editing/pasteboard/paste-blockquote-and-paragraph-break-expected.txt

We copy and paste a blockquoted paragraph plus a paragraph break.
The paragraph break shouldn't be inside the blockquote on paste.
You should see 'hello' (blockquoted), 'world' (not quoted, black text), 'hello' (blockquoted), empty paragraph.
See <rdar://problem/5368833<

Before paste:
| <#selection-caret>
| <br>
| <blockquote>
|   id="blockquote"
|   type="cite"
|   "hello"
|   <br>
| <br>

After paste:
| <blockquote>
|   id="blockquote"
|   type="cite"
|   "hello"
|   <br>
| <div>
|   <#selection-caret>
|   <br>
| <blockquote>
|   id="blockquote"
|   type="cite"
|   "hello"
|   <br>
| <br>

After inserting "world":
| <blockquote>
|   id="blockquote"
|   type="cite"
|   "hello"
|   <br>
| <div>
|   "world<#selection-caret>"
|   <br>
| <blockquote>
|   id="blockquote"
|   type="cite"
|   "hello"
|   <br>
| <br>
